refactor: update vue & adjust file dir
