+++ package.json
@@ -10,16 +10,16 @@
   "dependencies": {
     "axios": "^0.21.1",
     "swiper": "^6.8.2",
-    "vue": "^3.0.5",
+    "vue": "^3.2.6",
     "vue-router": "^4.0.11"
   },
   "devDependencies": {
     "@types/node": "^16.7.4",
-    "@vitejs/plugin-vue": "^1.3.0",
-    "@vue/compiler-sfc": "^3.0.5",
+    "@vitejs/plugin-vue": "^1.6.0",
+    "@vue/compiler-sfc": "^3.2.6",
     "less": "^4.1.1",
     "typescript": "^4.3.2",
-    "vite": "^2.4.4",
-    "vue-tsc": "^0.2.2"
+    "vite": "^2.5.2",
+    "vue-tsc": "^0.2.3"
   }
 }
