+++ src/App.vue
@@ -4,13 +4,8 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import Home from '@/views/Home.vue'
-
 export default defineComponent({
-  name: 'App',
-  components: {
-    Home
-  }
+  name: 'App'
 })
 </script>
 
@@ -19,9 +14,5 @@ export default defineComponent({
   margin: 0;
   padding: 0;
   box-sizing: border-box;
-}
-
-.xt-icon {
-  width: .5rem;
 }
 </style>
