--- src/components/Temp.vue
@@ -1,13 +1,0 @@
-<template></template>
-
-<script lang="ts">
-import { defineComponent } from 'vue'
-export default defineComponent({
-    name: 'Temp',
-    setup: () => {
-        return {}
-    }
-})
-</script>
-
-<style lang="less" scoped></style>
