+++ src/router/index.ts
@@ -1,7 +1,7 @@
 import { createRouter, createWebHashHistory } from "vue-router";
 
-const Home = () => import('@/views/Home.vue');
-const Songs = () => import('@/views/Songs.vue');
+const Home = () => import('@/views/home/index.vue');
+const Songs = () => import('@/views/songs/index.vue');
 
 
 const routes = [
