--- src/views/Songs.vue
@@ -1,15 +1,0 @@
-<template></template>
-
-<script lang="ts">
-import { defineComponent } from 'vue'
-import { useRoute } from 'vue-router'
-
-export default defineComponent({
-    name: 'Songs',
-    setup: () => {
-        return {}
-    }
-})
-</script>
-
-<style lang="less" scoped></style>
