+++ src/views/songs/components/ListViewSongs.vue
@@ -1,0 +1,13 @@
+<template>
+    <div>
+        这是一些歌
+    </div>
+</template>
+
+<script lang="ts" setup>
+
+</script>
+
+<style scoped>
+
+</style>
