+++ src/views/songs/components/ListViewTop.vue
@@ -1,0 +1,17 @@
+<template>
+    <div class="list-view-top">
+        <div class="left">返回</div>
+        <div class="right"></div>
+    </div>
+    <div class="list-view-info">
+        这是作者信息
+    </div>
+</template>
+
+<script lang="ts" setup>
+
+</script>
+
+<style scoped>
+
+</style>
